backend/uagents_system/requirements.txt
# HypeLens uAgents System Requirements
# =====================================
# Install with: pip install -r requirements.txt

# Fetch.ai uAgents Framework
uagents>=0.13.0

# Already in parent requirements.txt but needed for standalone runs
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
pinecone>=3.0.0
requests>=2.31.0
